llmserver/requirements.txt
grpcio==1.73.0
grpcio-tools==1.71.0
protobuf==5.29.5
langchain==0.3.26
langchain-google-genai==2.0.10
python-dotenv==1.1.1